VoicerGlobalControl: Change GUI to use (new) LayoutValueSlider
Previously, the Voicer GUI used crucial-library NumberEditors
for global control sliders. Something changed (probably view
sizing in NumberEditor), causing an ugly display. This commit
cuts out the external dependency and replaces NumberEditors
with a view (now in ddwGUIEnhancements) that guarantees the
number box and slider will always be in a horizontal row.